refactor(flow): move gate config resolution into postgres-approval-gate façade (BR32 Lot 4 Slice 1 Step 1)
