+++ BRANCH.md
@@ -51,6 +51,7 @@ Extract `todo-orchestration.ts` + `queue-manager.ts` + `default-workflows.ts` + 
   - Mirror the exception in this file under `## Feedback Loop`.
 
 ## Feedback Loop
+- **BR26-Q1** (Lot 4 Slice 1, status: `deferred`): tightening `PostgresApprovalGate.signal(runId, decision)` to consume the `decision` argument (gate-aware approval/rejection branching) would change observable behavior of `resumeRun` (today resume is unconditional after pause). BR-26 is contractually a **behavior-preserving extraction** (replay must stay byte-identical). `signal()` therefore stays advisory in Lot 4 Slice 1 — the boundary is locked at the type level (`GateDecision` accepted) but the parameter remains `_decision` until the gate-aware path lands in a follow-up branch (BR-27 or Lot N-3 under `BR26-EX3`). Recorded by Lot 4 Slice 1 sub-agent.
 - **BR26-EX5** (Lot 3, scope exception): `api/Dockerfile` is added to the touched paths. Reason: activating `@sentropic/flow` as an api workspace dependency (per `rules/architecture.md` "package extraction must be activated by real app consumption") requires (i) copying `packages/flow/package.json` into the base stage's pre-`npm ci` workspace shape and (ii) building `@sentropic/flow` after `COPY . .` so its `dist/index.d.ts` exists for the api `tsc --noEmit` typecheck. Mirrors the existing pattern for `@sentropic/llm-mesh`. Impact: 2 line additions in `api/Dockerfile`. Rollback: revert the diff if the package activation is rolled back.
 - **BR26-FB-01** (Lot 2, severity: `attention`, status: `resolved`): baseline `29b2c243` predates `origin/main` security fixes (`fix/security-high-vulnerabilities` merged at `edbe7d24`). The api Dockerfile audit gate (`npm audit --audit-level=high --omit=dev --workspaces --include-workspace-root`) fails on this branch as soon as any change invalidates the Docker layer cache, due to the Svelte `devalue` high-severity advisory (GHSA-77vg-94rm-hx3p). Lot 2 lockfile addition for `@sentropic/flow` triggers cache invalidation, surfacing the failure. Resolution path: rebase the branch onto `origin/main` (picks up vite/flatted/fast-xml-builder/svelte fixes) before Lot 3 or earlier. Recorded by Lot 2 sub-agent; deferred to conductor decision. **Resolved at Lot 3 by merge commit `4e9209cb` (Lot 3 baseline) which incorporates `origin/main` PR #152 + PR #157 (vite, flatted, fast-xml-builder, sveltekit/devalue HIGH advisories).**
